requirements.txt
# 核心框架
langchain==0.1.0
langchain-community==0.0.13

# 向量数据库和检索
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# 本地LLM推理
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
bitsandbytes==0.41.3  # 模型量化

# 文档处理
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.11.8

# Web界面
gradio==4.16.0

# 工具库
numpy==1.24.3
pyyaml==6.0.1
tqdm==4.66.1
openai
ragas
datasets
langchain-openai
